requirements.txt
earthengine-api>=0.1.370
google-auth>=2.0
google-api-python-client>=2.0
google-auth-httplib2>=0.2
numpy>=1.24
pandas>=2.0
geopandas>=0.14
rasterio>=1.3
pyproj>=3.6
scipy>=1.11
matplotlib>=3.8
shapely>=2.0
cartopy>=0.22
